backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-limiter==3.7.0
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.1.1
pytest-flask==1.3.0
flasgger==0.9.7.1
ruff==0.3.3
